requirements.txt
playwright
beautifulsoup4
python-dotenv
fastapi
uvicorn[standard]
pydantic